organize initial version of project
